made thresshold value a float

+++ lib/frame_detector_impl.h
@@ -228,7 +228,7 @@ public:
    * @param threshold : threshold value to use
    */
   frame_detector_impl(uint8_t sf, uint32_t samp_rate, uint32_t bw,
-                      uint32_t threshold);
+                      float threshold);
 
   /**
    * @brief Destroy the frame detector impl object
